+++ src/extension.ts
@@ -1,23 +1,31 @@
 import * as vscode from 'vscode';
 import { GuidCodeLensProvider } from "./GuidCodeLensProvider";
 import { GuidCompletionItemProvider } from "./GuidCompletionItemProvider";
+import { GuidHoverProvider } from './GuidHoverProvider';
 
 let disposables: vscode.Disposable[] = [];
 
 export function activate(context: vscode.ExtensionContext)
 {
-	const guidCompletionItemProvider = new GuidCompletionItemProvider();
-	const guidCodeLensProvider = new GuidCodeLensProvider();
-
-	const provider = vscode.languages.registerCompletionItemProvider(
+	const registeredCompletionItemProvider = vscode.languages.registerCompletionItemProvider(
 		"markdown",
-		guidCompletionItemProvider,
+		new GuidCompletionItemProvider(),
 		"@"
 	);
 
-	context.subscriptions.push(provider);
+	const registeredCodeLensProvider = vscode.languages.registerCodeLensProvider(
+		"markdown", 
+		new GuidCodeLensProvider()
+	);
 
-	vscode.languages.registerCodeLensProvider("markdown", guidCodeLensProvider);
+	const registeredHoverProvider = vscode.languages.registerHoverProvider(
+		"markdown",
+		new GuidHoverProvider()
+	);
+
+	context.subscriptions.push(registeredCompletionItemProvider);
+	context.subscriptions.push(registeredCodeLensProvider);
+	context.subscriptions.push(registeredHoverProvider);
 }
 
 export function deactivate()