+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 
-- None
+- Add hover provider to show the name if a guid is hovered.
 
 ## 0.0.3